refactor: shuffling around some more; non-functioning WIP
[travis skip]
